requirements.txt
tensorflow
numpy~=1.26.2
scipy~=1.10.1
pandas~=2.0.3
mne~=1.6.1
werkzeug~=3.0.2
flask~=3.0.3
gunicorn

keras~=2.12.0
imblearn~=0.0
sklearn~=0.0.post7
scikit-learn~=1.3.2
pyEDFlib~=0.1.37
tqdm~=4.64.1
seaborn~=0.13.0
matplotlib~=3.7.2
paho-mqtt~=1.6.1
requests~=2.31.0